Fix casts and constants

+++ src/event_handlers/websocket/local/_InvalidMessage.cpp
@@ -14,5 +14,5 @@ void _Private::HandleInvalidMessage(std::uint8_t socketId, const OpenShock::Seri
     return;
   }
 
-  ESP_LOGE(TAG, "Invalid message type: %d", root->payload_type());
+  ESP_LOGE(TAG, "Invalid message type: %hhu", static_cast<std::uint8_t>(root->payload_type()));
 }
